refactor: API 컨벤션 통일
- API 파일이름 수정
- api.ts 삭제
