+++ src/apis/mailBox.ts
@@ -1,8 +1,8 @@
-import { getRequest, postRequest } from './api';
+import client from './client';
 
 export const getMailbox = async () => {
   try {
-    const response = getRequest('/api/mailbox');
+    const response = client.get('/api/mailbox');
     if (!response) throw new Error('error while fetching mailbox data');
     return response;
   } catch (error) {
@@ -12,7 +12,7 @@ export const getMailbox = async () => {
 
 export const getMailboxDetail = async (id: number) => {
   try {
-    const response = getRequest(`/api/mailbox/${id}`);
+    const response = client.get(`/api/mailbox/${id}`);
     if (!response) throw new Error('error while fetching mailbox detail data');
     return response;
   } catch (error) {
@@ -22,7 +22,7 @@ export const getMailboxDetail = async (id: number) => {
 
 export const postMailboxDisconnect = async (id: number) => {
   try {
-    const response = postRequest(`/api/mailbox/${id}/disconnect`);
+    const response = client.post(`/api/mailbox/${id}/disconnect`);
     if (!response) throw new Error('error while disconnecting mailbox');
     return response;
   } catch (error) {
